requirements.txt
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
requests==2.31.0
python-dotenv==1.0.0
ruff==0.1.9
httpx==0.25.2
faker==22.0.0
pyyaml==6.0.1
